refactor: 각 모델별 폴더 분리
